REFACTOR: move socket controller to other package
